+++ web/modules/custom/ai_screening_project_track/src/Status.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace Drupal\ai_screening_project_track;
+
+use Drupal\Core\StringTranslation\TranslatableMarkup;
+
+/**
+ * Project track/tool status.
+ */
+enum Status: string {
+  case NONE = '';
+  case NEW = 'new';
+  case IN_PROGRESS = 'in_progress';
+  case COMPLETED = 'completed';
+
+  /**
+   * Get status options as key/value pair.
+   */
+  public static function asOptions(): array {
+    return [
+      self::NONE->value => new TranslatableMarkup('None'),
+      self::NEW->value => new TranslatableMarkup('New'),
+      self::IN_PROGRESS->value => new TranslatableMarkup('In progress'),
+      self::COMPLETED->value => new TranslatableMarkup('Completed'),
+    ];
+  }
+
+  /**
+   * Translate a project track/tool status.
+   */
+  public function getTranslatable(): TranslatableMarkup {
+    return static::asOptions()[$this->value];
+  }
+
+}