+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
 
+- [PR-42](https://github.com/itk-dev/ai-screening/pull/42)
+  - Added theming for project track.
+  - Added Enum for Evaluation values.
+  - Opdated status Enum
+  - Updated form for project track
 - [PR-41](https://github.com/itk-dev/ai-screening/pull/41)
   Remove include of site-search
 - [PR-36](https://github.com/itk-dev/ai-screening/pull/36)